requirements.txt
w1thermsensor==1.1.2
PyYAML==5.2
pyjwt==1.7.1
paho-mqtt==1.4.0
cryptography==2.7

#flaky==3.6.1
#gcp-devrel-py-tools==0.0.15
#google-api-python-client==1.7.11
#google-auth-httplib2==0.0.3
#google-auth==1.6.3
#google-cloud-pubsub==1.0.0
#google-cloud-iot==0.3.0
#grpc-google-iam-v1==0.12.3